refactor: refactor and update plugin-file
