+++ src/index.ts
@@ -1,5 +1,2 @@
-export {
-  default as FileItemPlugin,
-  GraaspPluginFileItemOptions,
-} from "./plugin";
-export * from "./utils/helpers";
+export { default as FileItemPlugin } from "./plugin";
+export { GraaspPluginFileItemOptions } from "./types";
